bring back trie and triehash benches
