+++ ethcore/Cargo.toml
@@ -70,6 +70,9 @@ unexpected = { path = "../util/unexpected" }
 journaldb = { path = "../util/journaldb" }
 tempdir = "0.3"
 
+[dev-dependencies]
+trie-standardmap = { path = "../util/trie-standardmap" }
+
 [features]
 jit = ["evm/jit"]
 evm-debug = ["slow-blocks"]
